refactor(domain/instrument): move `instrument` to `instrument.model` namespace
Restructured `instrument` package to `instrument.model` for better modularity and clarity. Updated imports across all affected modules accordingly.
