+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/contract/spot/catalog/persistence/jpa/FxSpotEntityMapper.java
@@ -2,7 +2,7 @@ package com.alligator.market.backend.instrument.asset.forex.contract.spot.catalo
 
 import com.alligator.market.backend.instrument.asset.forex.reference.currency.catalog.persistence.jpa.CurrencyEntity;
 import com.alligator.market.backend.instrument.asset.forex.reference.currency.catalog.persistence.jpa.CurrencyEntityMapper;
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpot;
 
 import java.util.Objects;
