+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/contract/spot/catalog/service/FxSpotCatalogService.java
@@ -1,7 +1,7 @@
 package com.alligator.market.backend.instrument.asset.forex.contract.spot.catalog.service;
 
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpot;
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
 
 import java.util.List;
