+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/contract/spot/catalog/service/FxSpotCatalogServiceImpl.java
@@ -3,7 +3,7 @@ package com.alligator.market.backend.instrument.asset.forex.contract.spot.catalo
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.exception.FxSpotNotFoundException;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpot;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.repository.FxSpotRepository;
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
