+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/contract/spot/catalog/service/FxSpotDomainFactory.java
@@ -2,7 +2,7 @@ package com.alligator.market.backend.instrument.asset.forex.contract.spot.catalo
 
 import com.alligator.market.backend.instrument.asset.forex.contract.spot.catalog.web.dto.in.FxSpotCreateDto;
 import com.alligator.market.backend.instrument.asset.forex.contract.spot.catalog.web.dto.in.FxSpotUpdateDto;
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.exception.CurrencyNotFoundException;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.model.Currency;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.vo.CurrencyCode;
