+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/contract/spot/catalog/web/FxSpotController.java
@@ -8,7 +8,7 @@ import com.alligator.market.backend.instrument.asset.forex.contract.spot.catalog
 import com.alligator.market.backend.instrument.asset.forex.contract.spot.catalog.web.dto.in.FxSpotUpdateDto;
 import com.alligator.market.backend.instrument.asset.forex.contract.spot.catalog.web.dto.mapper.FxSpotDtoMapper;
 import com.alligator.market.backend.instrument.asset.forex.contract.spot.catalog.web.dto.out.FxSpotResponseDto;
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpot;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
