+++ backend/src/main/java/com/alligator/market/backend/instrument/catalog/persistence/jpa/converter/AssetClassConverter.java
@@ -1,6 +1,6 @@
 package com.alligator.market.backend.instrument.catalog.persistence.jpa.converter;
 
-import com.alligator.market.domain.instrument.type.AssetClass;
+import com.alligator.market.domain.instrument.model.AssetClass;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
