+++ backend/src/main/java/com/alligator/market/backend/provider/adapter/moex/iss/config/MoexIssProviderConfig.java
@@ -2,7 +2,7 @@ package com.alligator.market.backend.provider.adapter.moex.iss.config;
 
 import com.alligator.market.backend.provider.adapter.moex.iss.MoexIssProvider;
 import com.alligator.market.backend.provider.adapter.moex.iss.config.handlers.MoexIssHandlersConfig;
-import com.alligator.market.domain.instrument.Instrument;
+import com.alligator.market.domain.instrument.model.Instrument;
 import com.alligator.market.domain.provider.model.handler.AbstractInstrumentHandler;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
