+++ backend/src/main/java/com/alligator/market/backend/provider/adapter/moex/iss/config/handlers/MoexIssHandlersConfig.java
@@ -3,7 +3,7 @@ package com.alligator.market.backend.provider.adapter.moex.iss.config.handlers;
 import com.alligator.market.backend.provider.adapter.moex.iss.MoexIssProvider;
 import com.alligator.market.backend.provider.adapter.moex.iss.config.instrument.forex.spot.handler.MoexIssFxSpotHandlerConfig;
 import com.alligator.market.backend.provider.adapter.moex.iss.instrument.forex.spot.handler.MoexIssFxSpotHandler;
-import com.alligator.market.domain.instrument.Instrument;
+import com.alligator.market.domain.instrument.model.Instrument;
 import com.alligator.market.domain.provider.model.handler.AbstractInstrumentHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
