+++ backend/src/main/java/com/alligator/market/backend/provider/adapter/moex/iss/instrument/forex/spot/support/MoexIssFxSpotSupportCatalog.java
@@ -4,7 +4,7 @@ import com.alligator.market.domain.instrument.asset.forex.reference.currency.mod
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.vo.CurrencyCode;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpot;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpotTenor;
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
 
 import java.util.*;
