+++ backend/src/main/java/com/alligator/market/backend/quote/streaming/QuoteStreamingOrchestrator.java
@@ -1,6 +1,6 @@
 package com.alligator.market.backend.quote.streaming;
 
-import com.alligator.market.domain.instrument.Instrument;
+import com.alligator.market.domain.instrument.model.Instrument;
 import com.alligator.market.domain.provider.model.vo.ProviderCode;
 import com.alligator.market.domain.provider.model.MarketDataProvider;
 import com.alligator.market.domain.quote.feed.InstrumentProviderResolver;
