+++ domain/src/main/java/com/alligator/market/domain/instrument/asset/forex/contract/spot/codec/FxSpotCodec.java
@@ -1,7 +1,7 @@
 package com.alligator.market.domain.instrument.asset.forex.contract.spot.codec;
 
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
-import com.alligator.market.domain.instrument.vo.InstrumentSymbol;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentSymbol;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.vo.CurrencyCode;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpotTenor;
