+++ domain/src/main/java/com/alligator/market/domain/instrument/asset/forex/contract/spot/model/FxSpot.java
@@ -1,10 +1,10 @@
 package com.alligator.market.domain.instrument.asset.forex.contract.spot.model;
 
-import com.alligator.market.domain.instrument.type.AssetClass;
-import com.alligator.market.domain.instrument.type.ContractType;
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
-import com.alligator.market.domain.instrument.Instrument;
-import com.alligator.market.domain.instrument.vo.InstrumentSymbol;
+import com.alligator.market.domain.instrument.model.AssetClass;
+import com.alligator.market.domain.instrument.model.ContractType;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.Instrument;
+import com.alligator.market.domain.instrument.model.vo.InstrumentSymbol;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.model.Currency;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.codec.FxSpotCodec;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.exception.FxSpotSameCurrenciesException;
