+++ domain/src/main/java/com/alligator/market/domain/instrument/asset/forex/contract/spot/repository/FxSpotRepository.java
@@ -1,6 +1,6 @@
 package com.alligator.market.domain.instrument.asset.forex.contract.spot.repository;
 
-import com.alligator.market.domain.instrument.vo.InstrumentCode;
+import com.alligator.market.domain.instrument.model.vo.InstrumentCode;
 import com.alligator.market.domain.instrument.asset.forex.reference.currency.vo.CurrencyCode;
 import com.alligator.market.domain.instrument.asset.forex.contract.spot.model.FxSpot;
