+++ domain/src/main/java/com/alligator/market/domain/instrument/model/AssetClass.java
@@ -1,4 +1,4 @@
-package com.alligator.market.domain.instrument.type;
+package com.alligator.market.domain.instrument.model;
 
 /**
  * Класс актива, к которому относится финансовый инструмент.
